arch-arm: Define separate stats for instruction/data S1/S2 walks
At the moment the walks stat in the TableWalker describes the total
number of table walks performed. This makes sense as our current
page table walking model assumes one table walker instantiated per
translation type. So the disambiguation in the stats file comes
from the name of table walker
(eg. itb_walker.walks, dtb_walker.walks)

As we are shifting towards having a single hardware table walker, we
need to disambiguate not by the name of the owning SimObject, but by
the name of the stat itself

Change-Id: Ia4defb178a8404d8255c2536808e2c2ee2259c05
Signed-off-by: Giacomo Travaglini <giacomo.travaglini@arm.com>