arch-vega: Implement DS_READ_*_TR_B16/B8/B6/B4
These instructions read either B64 or B96 bits the same as the
non-transponse instructions according to the docs. They then transpose
the values 16-threads at a time and write to VGPRs. The docs indicate
that multiple transposes might be needed for certain MFMA sizes, in most
cases two. These instructions therefore transpose 16x16 16-bit values,
16x32 8-bit values, and 32x32 6-bit or 4-bit values.

As a simple test it was confirmed that transposing a set of values A,
which are meant to represent a matrx, return the same original values
when transposing the transposed values.