arch-riscv: Allow load-reserved accesses to use the TLB (#2331)
PR #1387 modified the TLB logic such that load-reserved memory access
commands to skip the TLB. This caused problems after the page table
walk. This change modifies the logic to allow them to use the TLB as
before #1387.

This change also refactors `special_access` variable to be a function in
the `MemAccessInfo` object to reduce code duplication.

Signed-off-by: Jason Lowe-Power <jlowepower@google.com>
Co-authored-by: Jason Lowe-Power <jlowepower@google.com>