stdlib: Append resource version to cached resource filename (#2368)
Prior to this change only 1 version of a particular resource could be
cached at anyone time due to setting the filename to the resource_id.
When a different version of an already cached resource was requested the
stdlib would delete the cached version to then download the requested
version.

This change permits a cache per version of a resource by setting the
filename to the resources_id + resource_version.