cpu: Add user-mode stats (#2308)
One major concern when running simulations, especially in FS mode, is
whether a simulation's ROI actually executes the benchmark of interest.
One major problem here is _kernel foo_: instructions executed by the
kernel that don't relate to the benchmark of interest.

This PR tries to make kernel foo more transparent by exposing the number
of instructions and ops committed in user and kernel mode. In most
cases, a simulation should ideally have few kernel-mode instructions
(say, <5%). This PR uses an existing method, `inUserMode()` in gem5's
ISA class, to determine if a thread is in user-mode every time it
commits an op. From this, we can track the number of user- and
kernel-mode instructions and ops committed by each thread.

**Some considerations:**
- Is `inUserMode()` reliable and correct?
- Do the calls to `getIsaPtr()` and `inUserMode()` harm performance?
This code is run on every op commit, so ideally we don't want to slow it
down. Alternatively, we could track each thread's mode as a boolean
variable and flip it when its mode changes.
- It would be nice to have an event similar to MAX_INSTS, but for
user-mode instructions only.

Thanks for taking a look at this, I hope it's helpful!