mem-ruby: return early response for software prefetches (#2311)
This patch adds support for early response of software prefetches for
ruby in a similar way as it's done in the classic cache, preventing
these instructions from stalling on the pipeline until the memory
request completes. This is done by sending an early response back to the
core just as the request arrives (leaving the instruction freedom to
commit), and pass through a copy whose response will be catched and
deleted before it's sent back to the core.