cpu: Define a serializing flag for RegClasses
This patch adds a serializing flag to RegClasses.

At the moment (and by default) this is effectively
identical to the isFixedMapping method: it calls
the isRenameable method which returns false for
InvalidRegClass and MiscRegClass registers.

So the meaning of the method is: can the following RegId
be read/written without serializing behaviour?

The reason why we add this "duplicate" is because we
want to change the behaviour for some MiscRegClass
registers. While the registers are generally speaking
non-renameable, we still want to provide the capability
to some of those to be read/written without serializing
execution, provided the register is properly tracked in
the scoreboard

Signed-off-by: Giacomo Travaglini <giacomo.travaglini.arm.com>
Change-Id: Ia53494208ec6e674ca4a433a1e40795ecd8e675b