stdlib: Wrong wiring of PTW ports with sequencers in CHI platform (#2606)
In the existing CHI PrivateL1CacheHierarchy the order of arguments
provided to the connect_walker_ports was reversed. The method binds the
first port to the I-side table walker and the second to the D-side table
walker.

In the aforementioned platform we are assigning the I-side sequencer
port to the D-side PTW and viceversa


Change-Id: Id03dafecc83837e483f0c196a37887da35faf715

Signed-off-by: Giacomo Travaglini <giacomo.travaglini.arm.com>
Co-authored-by: Giacomo Travaglini <giacomo.travaglini.arm.com>