tests: fix failing FIFO RP daily tests (#2646)
The unittests-fast-debug (fast) tests in the daily tests are failing due
to the FIFO replacement policy TouchNull death test. The assert that is
supposed to be thrown in the test doesn't occur, as the gem5.fast build
optimizes out assertions. This commit makes it so that the test is
skipped on fast builds.