arch-riscv: fix `vm*` for `VLEN/EEW < 8` and `LMUL > 1` (#2626)
As reported here: [#2622](https://github.com/gem5/gem5/issues/2622),
`vmseq` (and others `vm*` instructions) doesn't work properly when
`VLEN/EEW < 8` and `LMUL > 1`.

The bug is in the `VMaskMergeMicroInst::execute()` function, when
`VLEN/EEW < 8` the `if` is executed.
However the initial value of the variable `bit_cnt` is wrong (must be
0), so when `LMUL > 1`, the following elements are write in the wrong
position of `Vd`.