stdlib: Remove GPU ip_discovery.bin if exists (#2690)
This will prevent the gem5 device specific IP discovery from being
symlinked which will cause driver loading to fail and subsequently the
application and simulation.

The ip_discovery.bin may or may not exist depending on what the user has
installed on the disk image, so use the safe option and remove whatever
is there and replace with known working gem5 provided discovery files.