cpu-o3: Instant ROB squash (#2670)
The squashWidth limits the number of instructions squashed
within on cycle. In case of a full, large ROB this can incure tens
of cycles during which the rename stage is blocked. E.g. with
160 instructions in the ROB and squash width of 8 rename will be
blocked 20 cycles.
With this change users can choose not to set the squashWidth which
will trigger gem5 to squash all instructions immediately.