mem-ruby: Optionally limit max number of pending transactions
This patch is optionally limiting the number of pending
transactions issued by the TlmGenerator

It does so by defining a max_pending_tran OptionalParam,
and by changing the behaviour of the pendingTransactions
member variable.

pendingTransactions were never deleted, even if they completed.
This is because at the end of the suite every transaction
return status was checked in sequence. So pendingTransaction
was a misnomer; it actually refered to a map of scheduled
transactions with action callbacks.

We now remove a transaction from the pending list whenever
the transaction completes. This in the TLM generator world,
happens when a transaction does not have any action
callback left. In this way, the pendingTransaction map
can give us at any time the real number of outstanding
transactions, and we can limit the issue rate accordingly
within the tick method

Change-Id: Id2bdf85163a9650067c71be3319d95e0656dd4d2
Signed-off-by: Giacomo Travaglini <giacomo.travaglini.arm.com>
Reviewed-by: Andreas Sandberg <andreas.sandberg@arm.com>