mem-cache: fix segmentation fault caused by MSHR debug flag (#2824)
This commit fixes a segmentation fault that sometimes occurs when the
MSHR debug flag is passed. Previously, in Cache::serviceMSHRTargets, the
call to popTarget() would attempt to use a pointer that had already been
deleted in a DPRINTF for the MSHR debug flag. Now, popTarget() checks if
the pointer is null before attempting to use it.

Fixes https://github.com/gem5/gem5/issues/2693.